chore: Moved routes to their directory
