+++ page/src/routes/mappage.jsx
@@ -1,0 +1,30 @@
+import Filters from "components/Filters/Filters";
+import MapView from "components/MapView/MapView";
+import ViewSelector from "components/ViewSelector/ViewSelector";
+import YearSelector from "components/YearSelector/YearSelector";
+import { createSearchParams, useNavigate, useParams, useSearchParams } from "react-router-dom";
+
+export function MapPage() {
+    const {year} = useParams();
+    const navigate = useNavigate();
+    const [searchParams] = useSearchParams();
+
+    return (
+        <div className="dcaGrid">
+          <Filters/>
+          <div className="dcaContent">
+            <YearSelector
+              isMap={true}
+              year={parseInt(year, 10)}
+              onChange={year => {
+                navigate(`/${year}/map?${createSearchParams(searchParams)}`);
+              }}
+            />
+
+            <ViewSelector selected={'map'}/>
+
+            <MapView year={year} />
+        </div>
+      </div>
+    );
+}
